Change everything to use 'BigDecimal' instead of 'Long'

+++ src/main/kotlin/utility/PokeApi.kt
@@ -17,7 +17,7 @@ class PokeApi {
         val thinPokemon = ThinPokemon(
             id = pokemon.id,
             name = pokemon.name.capitalize(),
-            xp = pokemon.baseExperience.toLong(),
+            xp = pokemon.baseExperience.toBigDecimal(),
             sprite = pokemon.sprites.frontDefault ?: pokemon.sprites.frontShiny
         )
